1. Simplify the model, Only two class: --Agent class for G/S/T --Landscape class for 4-state landscape; revise the way to generate landscape, binary first and then quaternary
2. Agents all search on the finest landscape, just perceiving it from different perspectives.
--adopt the default element design, resonating with the mindset and cognitive fixation; perhaps removed.
--adopt the shelter "*" element design, resonating with the myopia induced ambiguity.

3. Introduce the different sub-area in the landscape, removed.

4. Adjust the ambiguity design
--the average pooling algorithm now is much more informative and consistent with neighborhood perception. removed
--use the binary-to-quaternary diverging to implement the average pooling reversely

+++ SharingWhen/EarlySharing/S_send_to_S.py
@@ -26,7 +26,7 @@ def func(N=None, K=None, state_num=None, generalist_expertise=None, specialist_e
     crowd = Crowd(N=N, agent_num=agent_num, landscape=landscape, state_num=state_num,
                            generalist_expertise=0, specialist_expertise=20)
     for agent in crowd.agents:
-        for _ in range(0.1 * search_iteration):
+        for _ in range(int(0.1 * search_iteration)):
             agent.search()
     for agent_index in range(agent_num):
         specialist = Agent(N=N, landscape=landscape, state_num=state_num, crowd=crowd,
